requirements.txt
requests>=2.31.0
click>=8.1.7
python-dotenv>=1.0.0
PyYAML>=6.0.1
Flask>=3.0.0
authlib>=1.3.0
tenacity>=8.2.3
rich>=13.7.0
slack-sdk>=3.27.0
